+++ src/libpgmoneta/prometheus.c
@@ -3121,8 +3121,8 @@ metrics_cache_invalidate(void)
 static bool
 metrics_cache_append(char* data)
 {
-   int origin_length = 0;
-   int append_length = 0;
+   size_t origin_length = 0;
+   size_t append_length = 0;
    struct prometheus_cache* cache;
 
    cache = (struct prometheus_cache*)prometheus_cache_shmem;
